feat(regional): add via support
