+++ src/client/Common/Storage.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-process-env */
 import Cookies from 'universal-cookie';
 import StorageInterface from 'shared/hooks/useStorage/StorageInterface';
 
@@ -5,7 +6,7 @@ const setCookieOptions = {
   expires: new Date('2037-12-12'),
   httpOnly: false,
   path: '/',
-  secure: true,
+  secure: process.env.NODE_ENV !== 'test',
   sameSite: 'strict' as 'strict',
 };
